rename & fix global configuration
