+++ extension/src/flow-cli/cli-provider.ts
@@ -3,23 +3,23 @@ import { StateCache } from '../utils/state-cache'
 import * as vscode from 'vscode'
 import { Settings } from '../settings/settings'
 import { isEqual } from 'lodash'
-import { CliBinary, BinaryVersionsProvider, KNOWN_FLOW_COMMANDS } from './binary-versions-provider'
+import { CliBinary, CliVersionsProvider, KNOWN_FLOW_COMMANDS } from './cli-versions-provider'
 
 export class CliProvider {
   #selectedBinaryName: BehaviorSubject<string>
   #currentBinary$: StateCache<CliBinary | null>
-  #binaryVersionsProvider: BinaryVersionsProvider
+  #cliVersionsProvider: CliVersionsProvider
   #settings: Settings
 
   constructor (settings: Settings) {
     const initialBinaryPath = settings.getSettings().flowCommand
 
     this.#settings = settings
-    this.#binaryVersionsProvider = new BinaryVersionsProvider([initialBinaryPath])
+    this.#cliVersionsProvider = new CliVersionsProvider([initialBinaryPath])
     this.#selectedBinaryName = new BehaviorSubject<string>(initialBinaryPath)
     this.#currentBinary$ = new StateCache(async () => {
       const name: string = this.#selectedBinaryName.getValue()
-      const versionCache = this.#binaryVersionsProvider.get(name)
+      const versionCache = this.#cliVersionsProvider.get(name)
       if (versionCache == null) return null
       return await versionCache.getValue()
     })
@@ -43,10 +43,10 @@ export class CliProvider {
     // Subscribe to changes in the selected binary to update the caches
     this.#selectedBinaryName.pipe(distinctUntilChanged(), startWith(null), pairwise()).subscribe(([prev, curr]) => {
       // Remove the previous binary from the cache
-      if (prev != null) this.#binaryVersionsProvider.remove(prev)
+      if (prev != null) this.#cliVersionsProvider.remove(prev)
 
       // Add the current binary to the cache
-      if (curr != null) this.#binaryVersionsProvider.add(curr)
+      if (curr != null) this.#cliVersionsProvider.add(curr)
 
       // Invalidate the current binary cache
       this.#currentBinary$.invalidate()
@@ -58,7 +58,11 @@ export class CliProvider {
   }
 
   async setCurrentBinary (name: string): Promise<void> {
-    await this.#settings.updateSettings({ flowCommand: name })
+    if (vscode.workspace.workspaceFolders == null) {
+      await this.#settings.updateSettings({ flowCommand: name }, vscode.ConfigurationTarget.Global)
+    } else {
+      await this.#settings.updateSettings({ flowCommand: name })
+    }
   }
 
   get currentBinary$ (): Observable<CliBinary | null> {
@@ -66,15 +70,15 @@ export class CliProvider {
   }
 
   async getBinaryVersions (): Promise<CliBinary[]> {
-    return await this.#binaryVersionsProvider.getVersions()
+    return await this.#cliVersionsProvider.getVersions()
   }
 
   get binaryVersions$ (): Observable<CliBinary[]> {
-    return this.#binaryVersionsProvider.versions$.pipe(distinctUntilChanged(isEqual))
+    return this.#cliVersionsProvider.versions$.pipe(distinctUntilChanged(isEqual))
   }
 
   // Refresh all cached binary versions
   refresh (): void {
-    this.#binaryVersionsProvider.refresh()
+    this.#cliVersionsProvider.refresh()
   }
 }
