+++ extension/src/server/language-server.ts
@@ -7,7 +7,7 @@ import { BehaviorSubject, Subscription, filter, firstValueFrom, skip } from 'rxj
 import { envVars } from '../utils/shell/env-vars'
 import { FlowConfig } from './flow-config'
 import { CliProvider } from '../flow-cli/cli-provider'
-import { KNOWN_FLOW_COMMANDS } from '../flow-cli/binary-versions-provider'
+import { KNOWN_FLOW_COMMANDS } from '../flow-cli/cli-versions-provider'
 
 // Identities for commands handled by the Language server
 const RELOAD_CONFIGURATION = 'cadence.server.flow.reloadConfiguration'
